AB#104721:Change polling interval from double to int (#196)

+++ app/HutchAgent/Config/WatchFolderOptions.cs
@@ -3,5 +3,5 @@ namespace HutchAgent.Config;
 public class WatchFolderOptions
 {
   public string Path { get; set; } = string.Empty;
-  public double PollingIntervalSeconds { get; set; } = 5.0;
+  public int PollingIntervalSeconds { get; set; } = 5;
 }
